+++ src/http_file/test.http
@@ -1,5 +1,5 @@
 POST https://jsonplaceholder.typicode.com/posts
-Content-Type application/json
+Content-Type: application/json
 
 {
     "title": "foo",
